Add files via upload
In my journey through data exploration and analysis, Power BI has been my trusted companion, enabling me to create a dynamic tapestry of visualizations. From harnessing the power of Power Queries to seamlessly merge and shape diverse datasets, to establishing intricate relationships and joining that unveil hidden connections, every dashboard and report is a testament to the art of data storytelling. With the flexibility and precision of Power BI, I've transformed raw data into actionable insights, providing a panoramic view of business landscapes and illuminating the path to informed decision-making.